TracMasterTickets 4.0.4dev: Allow image format to be specified
The first entry in `[mastertickets]` `acceptable_formats`
will be used by default. The PNG format will be used if the
list is empty. The new default value for `acceptable_formats`
is `svg,png,cmapx`, so SVG will be the default format
if `acceptable_formats` is not specified in the
`[mastertickets]` section of trac.ini.

Fixes #13631.